Isolated Stats and JMX specific code
