+++ pkg/collector/corechecks/loader.go
@@ -57,7 +57,7 @@ func (gl *GoCheckLoader) Load(config check.Config) ([]check.Check, error) {
 	checks := []check.Check{}
 
 	// If JMX check, just skip - coincidence
-	if check.IsConfigJMX(config.Name, config.InitConfig) {
+	if check.IsJMXConfig(config.Name, config.InitConfig) {
 		return checks, fmt.Errorf("check %s appears to be a JMX check - skipping", config.Name)
 	}
